requirements.txt
streamlit
langchain
langchain-community
langgraph
python-dotenv
openai